backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
spacy==3.8.9
python-multipart==0.0.6
datasets
textblob==0.17.1
torch==2.6.0
python-docx==1.0.1
PyPDF2==3.0.1
